Fix bug in exclusive policy where it would crash/hang when subpools were scaled to 0.
Adds support for explicit shutdown of pools. This includes
switching from the .terminate() method I was using on workers (not
actually sure what that was, I can't find any docs for it or even a
reference in the Ray source). Now I use ray.actor.exit_actor() which is
the official documented way to shutdown an actor.

Adds a new stress test to poolTest.py to test for these behaviors.