Test and fix exclusive policy. From an API perspecitve, this added a 'shutdown' method to pools which isn't strictly necessary but helps with testing and accounting.
Internally, this involved a few minor bugs and a fairly large change to
the balance policy to allow for a 'force' option that issues requests to
busy workers so that none ever pile up inside the policy. This was
needed because the exclusive policy might scale groups down, preventing
pending requests from ever running. Regular BalancePolicy doesn't use
force because it's bad for load-balancing, but Exclusive needs to. I'm
still not totally wild about this, but I'm not sure how else to deal
with the deadlock.